+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### **Added**
+
 - adding new `integration/opensearch-tunnel` module
+- Added unit-tests, enabled logging and versioning on buckets for datalake-buckets module
+- Added Unit tests to metadata-storage module
+
 
 ### **Changed**
 
@@ -18,6 +22,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored Rosbag Image pipeline module to include scene detection workflow in the DAG and standardized inputs consumption from external modules
 - Added manifest files supporting AV/ADAS as AWS-Solutions
 - updating `post-processing/yolo-object-detection` with license and tests
+- syncing `integration/opensearch-tunnel` module
+- CDK version change, changed license-headers on datalake-buckets module
+- CDK Version change, sonar checks improvements and changed license-headers on metadata-storage module
+- upgraded CDK version, changed the license-headers, deployed sagemaker jobs inside VPC sonar code improvements on rosbag-image-pipeline module
+
+
 
 ### **Removed**